shill: dbus-client: Update default device change callback
Prior to this patch, the default device callback is triggered on
changes of:
1) default device state,
2) default service connected / disconnected,
3) default device of the default service,
4) default device property.

With (2), a device change of [a] -> [b] would be seen as [a] -> [],
followed by [] -> [b]. This is not fully correct as there would be
a slight unwanted disconnection timing when default device is empty.
This can be observed when DNS proxy is changing from VPN to the
physical network.

We can remove (2) by adding a empty device callback whenever the
default service is updated to be "" or "/". This is because alongside
a default service disconnection, we'd always see the default device
go away as well. The same is true for the disconnected -> connected
case.

While at it, this patch also updated the logging of (3) as it is very
uncommon (or maybe impossible?) for the device of a default service to
be changed without first updating the default service.

BUG=b:253545481
TEST=units
TEST=Switch between eth, wifi, vpn, no network;
     Observe correct callbacks;
TEST=tast run <DUT> network.DNSProxy*
TEST=tast run <DUT> network.ARCToHostVPN
     (with the precondition updated to arcBooted)

Change-Id: I44ec236840b99b482bc095f72841bcf20449d68c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4178300
Commit-Queue: Jason Jeremy Iman <jasongustaman@chromium.org>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Tested-by: Jason Jeremy Iman <jasongustaman@chromium.org>
Reviewed-by: Garrick Evans <garrick@chromium.org>