cryptohome: data_migrator: Allow arcvm-data-migrator to convert UID/GID
In virtio-fs /data, UID and GID of files in ARC's /data are translated
to other values on the host side, so they need to be converted back when
we migrate it to virtio-blk /data.
This CL therefore adds to the delegate class of MigrationHelper a method
to convert file metadata and skip the migration of the file on
conversion failure. Also overrides it in
ArcVmDataMigrationHelperDelegate to implement the UID/GID conversion and
adds related unit tests.

BUG=b:258402655, b:251764421
TEST=Manually run the migration and check that UID and GID of files are
TEST=translated.
TEST=FEATURES=test emerge-hatch cryptohome arcvm-data-migrator
TEST=tast run <DUT> cryptohome.VaultMigration.fscrypt_v1
TEST=tast run <DUT> cryptohome.VaultMigration.fscrypt_v2

Change-Id: Ied877d5c422b481da7b0b9dc4f5b4d854cf74035
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4183895
Reviewed-by: Youkichi Hosoi <youkichihosoi@chromium.org>
Tested-by: Momoko Hattori <momohatt@chromium.org>
Commit-Queue: Momoko Hattori <momohatt@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>