shill: cellular: use empty APN as fallback after the try list
When none of the items in the APN try list gives a valid connection,
the empty APN ("") is used as fallback, instead of attempting a
connect call without APN field, which is guaranteed to fail in
ModemManager with an InvalidArgs error.

The logic populating the KeyValueStore with the settings to use is
also updated so that the APN entry validity check (whether
SetApnProperties() returns true or false) is done on every attempt to
use a new element of the list. This was incorrectly only being checked
for the first item on the list.

BUG=b:248218738
TEST=cros_workon_make --board=${BOARD} --test shill

Change-Id: I520ab32607823d2a22de15637b110c6700e69b25
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4184225
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Tested-by: Aleksander Morgado <aleksandermj@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Aleksander Morgado <aleksandermj@google.com>