init: killers: Remove deprecated and unused term_process() function
Back in January of 2015 in <https://crrev.com/c/235804>, it was noted
that term_process() had some races in it and it was deprecated.

As far as I can tell, term_process() continued having at least one
user until at least May of 2017 when we landed
<https://crrev.com/c/502987>.

As of today, it appears that there are no users of this
function. Searching through code search finds no mentions other than
the definition in this file.

Let's delete it!

BUG=b:261373856
TEST=CQ

Change-Id: I76b0710256b1d4f50b454e2507cacd87858a7837
Disallow-Recycled-Builds: test-failures
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4185456
Reviewed-by: Brian Norris <briannorris@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>