metrics: Exempt usb_device project to always be recorded to the disk
Allow usb_device project to always be recorded to the disk disregarding
the UMA consent status. This is due to the known flakiness where
structured metrics may not be properly saved to the disk despite UMA
consent is on. Considering high priority of usb_device metrics, add it
to exemption so that we can collect metrics for sure. This exemption
shall be removed once the flakiness is fixed.

If UMA consent is off, this bypass would save usb_device metrics to the
disk, but Chromium Metrics Reader would not report it to Google servers.
The saved metrics are removed from the disk after 10 minutes.

BUG=b:265169949
TEST=FEATURES=test emerge-${BOARD} metrics
TEST=Check structured metrics recorded to the disk on a DUT

Change-Id: I1a286515ca5ce8534c8aa9d11047f8e933b73012
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4189872
Reviewed-by: Tony Yeoman <tby@chromium.org>
Tested-by: Won Chung <wonchung@google.com>