secanomalyd: adds known mount trackers to context
Adds the capability for remembering whether a known mount has persisted
across two consecutive scan intervals. Makes the SystemContext object a
Daemon class data member so that context can be preserved across scans.
Changes the context check function for /run/arc/shared_mounts/data mount
to consider if the mount was observed previously, and only report it as
known if it was not observed in the previous scan. It will also be
reported as known if the user is not logged in, as the mount is only
expected to be present in the logged in state. For more background on
this known mount please see b/219574442

BUG=b:242605262
TEST=When /run/arc/shared_mounts/data persists across two scan events,
it is not treated as a known mount and recorded.
TEST=When /run/arc/shared_mounts/data is present during a single scan,
it is considered a known mount and not recorded.
TEST=When the mount is observed on the login screen, it is never
considered known and is always recorded.
TEST=All secanomalyd unit and Tast tests pass.

Change-Id: Iba4aaf0ad1441e7909eceff67f72bdb9e30cf237
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4189874
Tested-by: Shervin Oloumi <enlightened@chromium.org>
Reviewed-by: Ben Scarlato <akhna@google.com>
Commit-Queue: Shervin Oloumi <enlightened@chromium.org>