dns-proxy: ARC proxies are not stopped on ARC shutdown
A router might implement a limit of IPv6 address per device. By
not stopping ARC proxies on ARC shutdown, we would be able to
reduce the number of IPv6 usage on CrOS.

The implementation is as follows:
- The proxies are only started on ARC startup.
- The proxies are removed when the underlying physical device is
  removed.
- The proxies are not stopped on ARC shutdown, but also won't be
  re-started if a proxy tracking the same device is already running.

We assumed that after a certain amount of time, the limit will be
relaxed by the router. By reducing the number of addresses the proxy
use, we are hoping that the previous addresses are released first
by the router before we reached the limit.

BUG=b:266496850
TEST=Restarting ARC; observe ARC DNS working;
TEST=Start ARC, remove / re-plug the physical device;
     Observe ARC DNS working;
TEST=Start ARC, start a physical device;
     Observe ARC DNS working;
TEST=tast run <DUT> network.DNSProxy.*

Change-Id: I01776916d49f592f406a6160236aa0cb9d7757d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4190917
Tested-by: Jason Jeremy Iman <jasongustaman@chromium.org>
Commit-Queue: Jason Jeremy Iman <jasongustaman@chromium.org>
Reviewed-by: Garrick Evans <garrick@chromium.org>