oobe_config: Use TPM encryption in rollback
Use TPM-based decryption on devices that support it, i.e. devices that
have a TPM 2.0 and the rollback space at 0x100e.

To migrate encryption, a second encrypted rollback data file is created.
It is called rollback_data_tpm and it is preserved across powerwash. It
contains TPM-encrypted rollback data. At the moment this file is only
created when requested via command-line flag, it can be used to test
that the decryption works. In the future, this file will always be
created during rollback.
See b/263065223 for details on the migration process.

The file called rollback_data contains openssl-encrypted data and is
still preserved across powerwash.

There are three versions of this code to think about:
1) Version pre-dating this CL, they only know about openssl encryption
   and rollback_data file. rollback_data_tpm is not preserved when
   rolling back to these old versions. As long as such versions are
   reachable via rollback, we do not create TPM-based encryption file.
2) The current version of this code. It creates both openssl and TPM
   encrypted data, if requested. It can decrypt both encryptions.
3) Future version of this code: It will create the openssl encrypted
   file on devices that do not have the rollback TPM space and create
   only the TPM encrypted file on devices that have the TPM space.

Note that there is also version (4) of the code which we can have once
all devices have the rollback space. Then we can delete rollback_data
file and openssl encryption. But that will only be in the far future
when all devices produced <=2022 are EOL.

BUG=b:233746744
TEST=0) FEATURES=test emerge-brya oobe_config
     1) USE="asan fuzzer" emerge-brya oobe_config
     2) Run manual rollback on devices with TPM2, TPM1.2, with and
        without rollback space.

Cq-Depend: 4092224
Change-Id: Id0aeb0d52a5f8ad1002d6884bd7110549616bfcf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4190201
Reviewed-by: Cristina Guerrero <crisguerrero@chromium.org>
Tested-by: Miriam Polzer <mpolzer@google.com>
Reviewed-by: Denis Kuznetsov <antrim@chromium.org>
Commit-Queue: Miriam Polzer <mpolzer@google.com>