shill: cellular: expose APN and APN type in CellularBearer
The PDN connection settings given to ModemManager during the
connection attempt will include the APN type set to "default".

It is ensured that ModemManager will expose in the DBus properties of
the Bearer object the same settings given in the connection attempt,
so the APN type will also be available in the Bearer object upon a
successful connection.

BUG=b:249387693
TEST=CQ passes

Change-Id: Iaef6aab33643da74bfe15d1c08949e4367329c30
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4199033
Tested-by: Aleksander Morgado <aleksandermj@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Aleksander Morgado <aleksandermj@google.com>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>