cryptohome: Drop GetKeyData implementation
Dependencies on this were gone as of the auth factor deployment so we
can drop the implementation as well. Unfortunately, this particular API
didn't actually dig very deeplying into the underlying implementation
(it just did a search of your VaultKeyset objects and read the KeyData
directly out of them) and so there isn't much secondary code that we get
to eliminate thanks to dropping this.

BUG=b:261089632
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I8e9f52fab12458104d4e20997af0ac0be8ea1989
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4195596
Reviewed-by: Betul Soysal <betuls@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>