biod: Fix maintenance routine trigger/timer
Fix the maintenance routine trigger/timer. Modify the maintenance timer
from repeating to one shot type. Add a check to the fingerprint sensor
mode before firing the maintenance mode. If the mode is not None,
reschedule the maintenance to 10 minutes in the future. Otherwise, run
the mainternance routine and schedule the next maintenance for the next
day.

BUG=b:187951992
TEST=cros_workon_make --board=hatch --test biod
TEST=Build ChromeOS image with ScheduleMaintenance every 30 seconds
     and RescheduleMaintenance every 5 seconds. Deploy on a hatch
     device. Try enrolling/authenticating and verify that the
     maintenance timer fired during those actions. Collect biod logs:
     http://shortn/_IW9WZC5BOL

Change-Id: I77800f89556d845035732a962bc0ddad28032699
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4193508
Tested-by: Firas Sammoura <fsammoura@google.com>
Commit-Queue: Firas Sammoura <fsammoura@google.com>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>