shill: vpn: Cache env vars passed from openvpn in driver
OpenVPNDriver::Notify() will be called in the cases of either connection
is set up initially (init), or the connection is reconnected (restart).
In the latter case, the environment variables passed in from openvpn
will not contain the routing information (see https://crbug.com/843407).

Currently, OpenVPNDriver will skip modifying fields related to routing,
if |inclusion_list| is not empty (which indicates routes have already
been set up in previous Notify() call). One corner case for this
solution is, in SetRoutes() we will also update the peer_address
property and add a route for the subnet, while they can be changed
before we return in ParseIPConfiguration(). This may not be a real case
but cause some inconsistency in the populated IPConfig::Properties
logically.

This patch changes the solution to keep the env vars dict instead of the
IPConfig::Properties, and recreate the IPConfig::Properties from the
updated dict every time Notify() is called, so that the fields in
IPConfig::Properties will be consistent.

Note that this solution actually assumes the env vars will not be unset
by intention (e.g., if a var no longer exists, it just means the value
is missed). The previous solution also has this assumption so nothing
should be affected.

BUG=b:262662614
TEST=FEATURES="test" emerge-${BOARD} shill
TEST=tast run $DUT network.VPNRouting.openvpn*
TEST=Connected to a OpenVPN, sent SIGUSR1 to the client process to
  trigger a restart, observed that in shill:
  - service was disconnected and then connected;
  - the dict passed in Notify() is not completed;
  - routes were set up correctly.

Change-Id: Ieda768ecd70d4ad7e0fa879dca2ecb3624bd37de
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4190900
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>