cryptohome: Add TPM-sim test for pwd update on recovery
Add an integration-style unit test for verifying AuthSession
implementation in this scenario: AuthenticateAuthFactor(recovery),
then UpdateAuthFactor(password). The test is parameterized to
check it regardless of whether the password is initially added
as VK or USS.

Note: the test doesn't verify what happens if we roll back to VK
afterwards. Per linked bug, we don't support authenticating in this
scenario anymore, but we should guarantee that the old password
isn't accepted anymore either, and add a test assertion for that -
it'll be done in a follow-up.

BUG=b:262632342,b:268104219
TEST=new unit test (cros_workon_make --test cryptohome)

Change-Id: I6201dcf77bb844f098f7fce39a43283c279db498
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4193162
Reviewed-by: Betul Soysal <betuls@google.com>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Maciek Slusarczyk <mslus@chromium.org>
Tested-by: Betul Soysal <betuls@google.com>
Commit-Queue: Betul Soysal <betuls@google.com>