u2fd: Use ListAuthFactors and not GetKeyData
The GetKeyData function is deprecated and to be removed soon, but is
still being used by u2fd to determine if a user has a PIN configured.
This rewrites that code to instead use the newer ListAuthFactors.

This usage didn't actually break when we enabled USS because (for now)
we're still also maintaining a parallel set of vault keysets. But once
we start to disable this will stop working correctly.

Plus we also just want to delete this operation from cryptohome.

BUG=b:261089632
TEST=u2fd unit tests (although they seem to have mediocre coverage)
TEST=ran tast u2fd tests

Change-Id: I3ddbf56c82050f9c5ed5cd096bacb890283ec102
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4194859
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Yi Chou <yich@google.com>
Reviewed-by: Andrey Pronin <apronin@chromium.org>