cryptohome: Move flags handling up to AuthSession::Create
This basically gets rid of the raw "flags" as an AuthSession constructor
parameter, replacing it with an is_ephemeral_user bool. The flags
handling probably should go even another layer higher, up to the
UserDataAuth layer, but that's a much more involved changes where this
was very quick and easy.

BUG=b:266704614
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I61185e510ea26e257806ffee9d504d5d2da9ba0c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4194870
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>