vm_tools: crostini_client: clarify DLC error message
The original DLC download was for Parallels, but we now use it for other
projects as well. Remove the "Parallels" mention from the error message.
The message already includes the name of the DLC, which should be
sufficient to figure out what is missing.

BUG=None
TEST=vmc start ... --tools-dlc termina-dlc ...

Change-Id: I95e51ea269bd2507af274e5961cf5b712a30b1c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4195598
Tested-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Abhishek Bhardwaj <abhishekbh@chromium.org>
Tested-by: Abhishek Bhardwaj <abhishekbh@chromium.org>