thinpool_migrator: Add stateful migrator class
The stateful migrator class corrals all the operations
required to migrate the stateful partition to use
an lvm2 thinpool. In order to do so, we need the
first 1M of the partition for lvm2 metadata. So, we

- resize the filesystem down the require amount, copy
over the partition header to the end of filesystem.

- generate thinpool metadata and write it out to the
offset in the newly created metadata space.

- generate lvm2 metadata and write it to the beginning
of the stateful partition.

- ... profit

BUG=b:232006113
TEST=unittests

Change-Id: I6f606599bf4190539b6917ce08a042381cd68fc3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4307565
Commit-Queue: Sarthak Kukreti <sarthakkukreti@google.com>
Tested-by: Sarthak Kukreti <sarthakkukreti@google.com>
Reviewed-by: Daniil Lunev <dlunev@chromium.org>