shill: network: separate Attach/DetachNetwork
This patch separates Service::SetAttachedNetwork into two different
functions for attaching and detaching a Network to/from a Service. At
the moment the usage in Device::SelectService guarantees that
SetAttachedNetwork is never called again for a Service if the Network
attached prior was not detached first. Separating both functions allow
to make this invariant clearer.

A similar refactoring is almost possible for Device::SelectService
itself. Analysis shows that SelectService is never called with an
existing selected Service for Ethernet and VPN. However the situation is
not as clear for WiFi and Cellular.

Separating the code path "select service" / "attach network" from the
code path "unselect service" / "detach network" will subsequently allow
to merge multiple layers of state transition into clearer transactions
for connecting or disconnecting a Service.

BUG=b:265607618
TEST=FEATURES=test emerge-$BOARD shill

Change-Id: I167e8bde8f37cbc2285400bfdce120ffe18e46e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4248062
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Tested-by: Hugo Benichi <hugobenichi@google.com>