private_computing: Add reading the status file from /var/lib
We discovered numerous membership check requests in the client
that are computationally intensive, despite our implementation
of reading from a preserved file. This suggests a potential bug
in the check membership conditions. To enhance performance,
we've added a double-check of the file in /var/lib. The client
updates the latest active date in this location after every
successful ping. If the file exists, we read it first; if not,
we fall back to reading from the preserved file.

BUG=chromium:1411103
TEST=cros_workon_make --board=${BOARD} ${PACKAGE_NAME} --test
     and flash to DUT and manually test on DUT

Change-Id: If64670ffb10e5a0bfdb68c8e173560af8342961a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4199696
Reviewed-by: Hirthanan Subenderan <hirthanan@google.com>
Tested-by: Eric Wang <qianwan@google.com>
Commit-Queue: Eric Wang <qianwan@google.com>