biod: Add unit tests for Session class destructor
Add unit tests for the destructor of the Session class for the cases
when the session object is valid (i.e. has a valid pointer) and invalid
(i.e. has an invalid pointer). Move the error field reset to outside the
conditional check of the validity of the Biometrics Manager pointer to
ensure it gets always cleared.

BUG=b:230478501
TEST=cros_workon_make --board=hatch --test biod

Change-Id: Icaf8ae4798e753585c7ecaa7aafacb3db3d017c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4199703
Reviewed-by: Bobby Casey <bobbycasey@google.com>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Commit-Queue: Firas Sammoura <fsammoura@google.com>
Tested-by: Firas Sammoura <fsammoura@google.com>