shill: cellular: Add is_required_by_carrier_spec to MODB
Add the property `is_required_by_carrier_spec` to the mobile operator
proto to enable carriers to define APNs that override any other APNs of
the same type. The main purpose of this value should be to enforce a DUN
APN on some carriers, preventing the user from using a different APN for
tethering.

BUG=b:266980208
TEST=USE="asan clang" FEATURES="test" emerge-amd64-generic shill

Change-Id: I00674f51fc308f84887a7ddb6841a3a364fc3c26
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4214317
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Aleksander Morgado <aleksandermj@google.com>