shill: cellular: new Cellular.PrimaryMultiplexedInterface property
This property for now exposes the same interface name as the main
Interface property name in the Device object in DBus.

When multiplexing is enabled, this property will expose the name of
the multiplexed network interface where IP config settings would
apply.

This property is only set right before the transition to Linked state,
and will be cleared when the cellular network is disconnected.

BUG=b:267073827
TEST=CQ passes
TEST=Connect and check /usr/local/lib/flimflam/test/list-devices
    ...
    Cellular.PrimaryMultiplexedInterface = wwan0
    ...
    Interface = wwan0
    Name = wwan0

Change-Id: I85517f02e8ad462e6edec5c301e3c3310ae9b9c2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4201436
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Tested-by: Aleksander Morgado <aleksandermj@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Aleksander Morgado <aleksandermj@google.com>