patchpanel: Install vpn rules for counters before routing
iptables rules for traffic counters should be installed before the rules
for routings, so that every packets sent to vpn can be counted. Since
counters rules do not affect the fate of a packet, this patch should not
affect routing except for delaying the routing setup for a little bit.

BUG=b:266169195
TEST=`tast run $DUT network.TrafficCounters.vpn` for 5 times

Change-Id: I82ade7878327d4425233feb694fba0ee26d57ec2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4202651
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Garrick Evans <garrick@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Jie Jiang <jiejiang@chromium.org>