vm_tools: 9s: allow readlinkat syscall
This is used when resolving the target of a symbolic link. In
particular, drivefs represents folder shortcuts in Google Drive
shared folders as a symlink, so listing a directory containing a folder
shortcut would cause the 9s server thread to be killed with SIGSYS.

Support for readlink on 9p shared directories was added in
https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4072744
and before that point, a 9p readlink request would just return EACCES,
so folder shortcuts would fail to list but would otherwise not cause
a failure.

With this patch applied, it is possbile to read the target of folder
shortcuts within Drive shared with Linux.

BUG=b:265877285
TEST=Create folder shortcut in Drive and ls -l in Crostini

Change-Id: I223a649d804d8ff3f2e69c15b80d621d5c8b2e4d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4204815
Reviewed-by: David Munro <davidmunro@google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Tested-by: Daniel Verkamp <dverkamp@chromium.org>