init: process_killer: Always log all processes that are killed
The process killer is never intended to be the primary way that
processes are killed. That means that even when it successfully kills
a process it already means we're in an error condition. Now that
process printing isn't so verbose, let's go ahead and print processes
whenever we try to kill them.

BUG=b:267198960
TEST=Logs include all killed processes.

Change-Id: I849d66aa78a19ec623778eba16785bc3c3569793
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4204824
Reviewed-by: Sarthak Kukreti <sarthakkukreti@google.com>
Commit-Queue: Douglas Anderson <dianders@chromium.org>
Tested-by: Sarthak Kukreti <sarthakkukreti@google.com>