init: chromeos_shutdown: Send some logs (incl process_killer) to /dev/kmsg
If something is going wrong with the chromeos_shutdown script it's
incredibly hard to figure out because the stdout/stderr don't go
anywhere. Let's instead redirect it to /dev/kmsg which means that the
output can be found in console-ramoops on the next reboot.

As part of this, let's tee the process_killer output to the same
place. A full justification added to comments in the
`chromeos_shutdown` script, but briefly this makes sense because:
* Even without this we see process terminations in the kernel ring
  buffer at reboot/shutdown time. Now we'll at least be able to tie
  together which ones were caused by process_killer.
* We want to reduce the number of processes that process_killer kills
  anyway (things should really be shut down before chromeos_shutdown
  runs), so this extra cruft in the log should call extra attention to
  this issue.

BUG=b:267198960, b:267224195
TEST=See output in /sys/fs/pstore/console-ramoops-0 after reboot
TEST=Confirm that /run/process_killer.log gets populated still

Change-Id: I340671c9921291184d9c3659f70ae79d84a489e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4206487
Tested-by: Douglas Anderson <dianders@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>