init: process_killer: Cowardly refuse to kill init
The `process_killer` program was never intended to be used to kill the
`init` program. While it _shouldn't_ happen, we are seeing an
unexpected and unexplained kill of the `init` process at around the
same time `process_killer` is running. Let's add an extra double-check
to make sure we don't kill `init`.

BUG=b:267198960, b:261373856
TEST=Run unittests

Change-Id: I45a2fc81ab7b6ea43d0baf7e49ae15017dea6843
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4204826
Tested-by: Sarthak Kukreti <sarthakkukreti@google.com>
Commit-Queue: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Sarthak Kukreti <sarthakkukreti@google.com>