u2fd: Use Blob for user_secret
In the past, we had used blobs instead of secure blobs for user_secret
because it's mainly used to separate secrets of different accounts on
the device. The main security of the credentials relies on the
auth-time-secret-hash and user presence. We've changed it to secure
blobs recently to be consistent with the other interfaces where
user_secret is passed as a secure blob, but loading too many secure
blobs will hit RLIMIT_MEMLOCK. This CL reverts that logic and use blobs
for user_secret instead to fix the crashes.

Note that with user_secrets as secure blobs, my DUT can only load <10
records without crashing.

BUG=b:265225402
TEST=u2fd unittests

Change-Id: Iaa5d23a9a573d19a9d08fb1d23cfdca646efccb0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4207538
Tested-by: Howard Yang <hcyang@google.com>
Reviewed-by: Yi Chou <yich@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>