missive: Don't create queue with corrupt file paths
This change returns an error status for the StorageQueue constructor if
a generation id is not found in any of the file paths in the directory
assigned to the StorageQueue. Previously we attempted to recover this
data using metadata files, but if the generation id is missing from every single data file path, it's safe to assume that the entire directory is corrupt and should be deleted.

BUG=b:267429525, b:267497955
TEST=FEATURES="test" emerge-${BOARD} missive

Change-Id: I7abe94b428e8484451c700568de2d4629ef41400
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4211444
Reviewed-by: Leonid Baraz <lbaraz@chromium.org>
Reviewed-by: Hong Xu <xuhong@google.com>
Tested-by: Josh Hilke <jrhilke@google.com>
Commit-Queue: Josh Hilke <jrhilke@google.com>