shill: cellular: Network lifecycle tied to the connected MM bearer
The Cellular object will no longer use the implicit Network object in
the parent Device. Instead, it will use a Network object that is
created as soon as ModemManager reports a successful connection.

A custom GetPrimaryNetwork() implementation is provided, which may
return nullptr if the modem is not in an active connected state. If
the modem is connected, the cellular specific Network object is
returned. This new Network object is tied exclusively to the Default
APN internet connection.

The unit tests are updated to ensure that whenever the modem is
assumed to be connected, a Network object is made available.

This is a first step in the integration of multiple network interfaces
in Cellular devices.

BUG=b:249386791
TEST=CQ passes

Change-Id: I985ee9cbeaf96e85d6845d3990f777a27e2dd2cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4215435
Tested-by: Aleksander Morgado <aleksandermj@google.com>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Commit-Queue: Aleksander Morgado <aleksandermj@google.com>