shill: cellular: allow connection on a multiplexed interface
The network interface reported in the Bearer object to setup a
connection does not need to be the same as the interface used to
instantiate the Cellular(Device) object.

This allows a Cellular(Device) object that was created for e.g. wwan0
be connected through a multiplexed mbimmux0.1 interface.

BUG=b:249386791
TEST=CQ passes

Change-Id: I3915dd3a6dd5841f1872acb836a8515ba7f5cc36
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4124194
Tested-by: Aleksander Morgado <aleksandermj@google.com>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Aleksander Morgado <aleksandermj@google.com>