shill: cellular: Add tethering_allowed to modb
Add the flag tethering_allowed to the mobile operator proto. The flag
can be used to restrict tethering on a carrier. The flag will start with
the value set to false to disable tethering in all carriers, and when
the feature is more mature, the default will be set to true.

BUG=b:249371954
TEST=Unit tests

Change-Id: Ib7f2e41eacb42fe1b7064d9f973b8edc104eae7d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4242419
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Aleksander Morgado <aleksandermj@google.com>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>