libbrillo: Add strong alias types or cryptohome Usernames
In cryptohome we make use of both usernames and hashed+salted usernames
in a bunch of internal code, with most of the basic manipulations being
provided by this library in libbrillo. One problem we've had is that it
can be very unclear which type of name you have in a variable or
parameter, because they're all just strings. We basically rely on
variable names to keep track of this but it's easy to make a mistake.

This change adds base::StrongAlias types for both username and the
hashed version, and overloads for all the standard function that use
them. The idea would be to convert the cryptohome code over to using the
new strong types and then get rid of the string overloads, although
this approach lets us do that incrementally instead of in one big
commit.

I'm open to other suggestions on the name of the sanitized type. I used
ObfuscatedUsername because that's what cryptohome uses internally fairly
consistently, but I think HashedUsername, SanitizedUsername, and
SaltedUsername could all work.

BUG=b:263367263
TEST=FEATURES=test emerge-$BOARD libbrillo cryptohome

Change-Id: I3d726c128959d6a8429c77288e1984680fd8c2f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4220317
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>