crash: Add ppid to crash_reporter-user-collection's ps log command
We're still seeing cases where the user_collector's early Chrome crash
mode is activating when it shouldn't, specifically, when chrome_crashpad
is already running.

Add parent process ID (ppid) to the list of things logged when
user-collector fails to collect (which happens when the early Chrome
crash mode activates erroneously). This will tell us if the problem is
that the crashpad process is being reparented earlier that I expected it
to be.

BUG=b:244787525
TEST=CQ

Change-Id: Ie750fa414cbb040adcc7d88893e9ff18c5c28edf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4221011
Commit-Queue: Ian Barkley-Yeung <iby@chromium.org>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Tested-by: Ian Barkley-Yeung <iby@chromium.org>