power: Avoid MetricsSender taking ownership of its dependencies
`MetricsSender` currently has a constructor which takes
a `std::unique_ptr<MetricLibraryInterface>`. While passing ownership is
convenient, it makes testing difficult because the calling code loses
control of the lifetime of any fake/mock dependencies it passes in. In
turn, this complicates upcoming CLs, which needs to interact more
heavily with its fake dependencies.

Update `MetricsSender` to just take a reference to its dependencies, and
leave ownership with the caller.

BUG=b:267701376
TEST=`cros_workon_make --board=$BOARD --test chromeos-base/power_manager`

Change-Id: Iee408cc6baaaa34ee1a485dbbfc24d4f6529f858
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4220464
Tested-by: David Greenaway <dgreenaway@google.com>
Reviewed-by: David Lattimore <dml@chromium.org>
Commit-Queue: David Greenaway <dgreenaway@google.com>