power: Simplify adaptive charging metrics test
Currently the tests have a significant amount of logic in it (such as
performing calculations of expected values, and concatenation of strings
to form expected metric names), making it a little unclear
to readers exactly what is being tested.

Simplify the tests, using more concrete inputs and outputs to make the
tests clearer.

BUG=b:267701376
TEST=`cros_workon_make --board=$BOARD --test chromeos-base/power_manager`

Change-Id: I919221f25ccee667ba1480e1e6d0de05f8a124b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4220466
Commit-Queue: David Greenaway <dgreenaway@google.com>
Reviewed-by: David Lattimore <dml@chromium.org>
Tested-by: David Greenaway <dgreenaway@google.com>