metrics: Add Clear method to FakeMetricsLibrary
Add a method `Clear` to `FakeMetricsLibrary` that removes all recorded
values from the fake.

This simplifies writing tests which have multiple "phases" of metrics
being sent: after a phase in a test is completed, the test can clear the
data in the fake. This means that the next phase of the test need not
worry about stale data or dealing with call counts from the previous
phase.

BUG=None
TEST=`cros_workon_make --board=$BOARD --test chromeos-base/metrics`

Change-Id: I999c824adb719fbc8a61d1151aa11d843ff16dc8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4220465
Auto-Submit: David Greenaway <dgreenaway@google.com>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Commit-Queue: David Greenaway <dgreenaway@google.com>
Tested-by: David Greenaway <dgreenaway@google.com>