cryptohome: Remove unused functions in KeysetManagement
This CL removes unused functions in KeysetMananagement class. In doin
so, the CL also rewrites some of the tests to use updated functions with
KeyBlobs.

Note: There is certainly more simplification that can be done with the
function calls in KeysetManagement, I've decided to reserve that for the
followup CL to only focus this CL on test rewrites due to removed
functions.

BUG=b:261022410
TEST=FEATURES=test emerge-zork cryptohome

Change-Id: Ic6f47d6e77f948b87bac6d682ee772311e3baeee
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4224613
Tested-by: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>