cryptohome: Use strong types for username
In order to make it much more difficult to pass the wrong type of
username into code, this rewrites as much of cryptohome to use the new
strong Username and ObfuscatedUsername types, instead of slinging these
values around as raw strings.

For the most part a lot of the code just works. Code which just passes
around these values doesn't significantly change, other than using
explicit types. There are some small changes to code on the boundaries,
that have to explicitly translate between the strong types and strings:

* getting usernames out of protos
* putting them into protos
* passing values into libhwsec calls (which don't use these types)
* turning values into filesystem paths

Most of these behaviors are fairly straightforward.

Updating tests is a bit more painful. One downside to the strong types
is that because they're std::string backed you can't make constexpr
values, so they're not suitable for global or static constants. In tests
which currently use `constexpr char` constants for these it's a bit
messy to update; in a lot of places I solved this by converting the
constants into const (not constexpr) non-static constants in the test
harness. This is a bit messy (that's not how one normally does "class
constants") but I don't think is too bad, and in a bunch of cases it
allowed me to not have to rewrite all the tests to manually construct
the type as much.

BUG=b:263367263
TEST=FEATURES=test emerge-$BOARD cryptohome
TEST=build_packages

Change-Id: Ib8ada2c4c17976a489f0e7257b5a074d76862ce0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4228352
Reviewed-by: Betul Soysal <betuls@google.com>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>